Minor CMakeLists tweaks on top of the tag CRYPTOPP_8_2_0
- Tag CRYPTOPP_8_2_0 corresponds to the cryptopp 8.2.0 release

- Builds in our environment, according to our conventions and
expectations.

- We only build STATIC libraries for third party things like this,
so the BUILD_SHARED code paths may not work properly with this
commit.

- Avoid "sh -c" chunks checking for CRYPTOPP_SOLARIS on MSVC builds
(obviously not necessary)

- Remove unnecessary forced include of winapifamily.h header file.
Builds just fine without it in our environment.

- Set OUTPUT_NAME property for MSVC builds

- Remove cryptest.exe target: was causing CMake build problems and
is not necessary.

This commit has some whitespace changes to unindent previously
conditional chunks. View diff with "no whitespace changes" to
see how simple the logic change is.